Introduce set-upstream-range
Refactor 'find' and introduce a new command 'set-upstream-range',
which, like 'find', sets upstream<>downstream links. The difference
is that while 'find' searches the commits of the symmetric difference
of the upstream and downstream, 'set-upstream-range' just looks at
each downstream commit in the given range. If the upstream link is a
valid commit then it will be linked. Validity is determined by
existence. Additionally, if an optional upstream range is given,
then the link must point to a commit within that range to be
considered valid.