Move -h help text to a man page
This allows 'git help abc' and man git-abc to work, which is the
more typical way for a git subcommand to have its documentation
displayed. It also gets the help text out of the script, which
is nice, as it was a bit annoying to scroll past. The disadvantages
are that we now need to maintain it (troff) and install it
separately.