Use git-sh-setup
This allows us to remove a few functions that did the same thing, as
well as better assimilate git-abc into the sh git script Borg.

We also risk changing the user interface a bit with this change, as
abc-todo-path must now be a directory, instead of a file. Likely
nobody was specifically configuring that anyway, and, as it must have
been a temporary file, then changing it to a [probably] temporary
directory, shouldn't be noticeable at all.