contrib/git-abc-backport-diff: Fix and improve
It turns out that git-range-diff was a bad choice for doing the
diff parsing since it's output is not stable. Switch to 'diff -u'
like the original git-backport-diff uses. Also, integrate with
git better by using git-sh-setup and implement git-backport-diff's
diff viewing feature (where we can use git-range-diff). Finally,
if the user has an ABC_NAMESPACE set in their environment, then
use that instead of a temporary one with a set-upstream-range.