Fix set-upstream-range with an upstream range input
We can't check that a commit C is in a range of commits M..N
by checking if M is an ancestor of C because if C came from a
merge then it won't be. Change the C in range test to

  C is an ancestor of N && C is not an ancestor of M

Also document the recommendation to use core.commitGraph in
order to improve the performance of range checking.