Improve input validation with range checking
set-upstream-range learns to check its range inputs and everything
else also learns to check that a given upstream or downstream commit
is in its expected range. This depends on core.commitGraph being
enabled though, as the checking would otherwise be too slow. Also,
while set-upstream, get-{upstream,downstream}, and flag now produce
errors when their inputs are not on the expected branches. We only
complain, but still proceed, when upstream links point to commits not
on expected branches.

We also introduce abc.range-checks to disable this checking even
when core.commitGraph is enabled. Sometimes we know we're safe and
just want more speed.