Introduce find hooks
Introduce three find hooks: abc-pre-find, abc-post-find, and
abc-ignore-evil-merge.  abc-pre-find is good for preparing the
upstream and downstream branches for 'find'. abc-post-find is good
for immediately rejecting candidate commits that are known to be
unwanted. And abc-ignore-evil-merge is good for eliminating merge
commits from the candidate list that are known to be safe.  For
example, if the "evil" changes are only whitespace.