Change range-checks default to depend on core.commitGraph
Range checking is quite expensive and in a large repository it doesn't
even make sense to try it without core.commitGraph enabled. Change
the default to depend on core.commitGraph and also improve the warnings
and documentation regarding it. And, when enabling core.commitGraph,
don't forget to create the initial commit-graph file with
'git commit-graph write'!