Check for unambiguous pattern in upstream link first
The 'cherry pick' pattern unambiguously identifies an upstream
link, while the 'commit' pattern may show up in a commit message
for other reasons. Checking for 'cherry pick' first, in case
there are both will ensure we get the linkage right when there's
a 'commit' pattern, which isn't the linkage, and a 'cherry pick'
pattern which is.