Pretty format placeholders may include colons
We don't want colons in the placeholders because we use colons to
split placeholders from regular expressions (see should-highlight
and upstream-link). Luckily colons may only be included in
placeholders when using one of the form %(...). In that case we
just use ): for our deliminator instead. Also, the %(...) placeholder
with a colon that we want to support at the moment is
%(trailers[:options]), which takes another placeholder modification
to ensure the regular expression is only applied to the trailer's
value. While at it, factor out the phs and re splitting to avoid
duplicate code.