Fix commit extraction
We already expected abc_extract_commit[s] to potentially return
more than one commit, since it looks at all lines of input, but
it should also return all the matching commits on the same line.
We get rid of the weird sed expression, which is nice, but end
up adding another pipe, which is not nice.