Add more command line options to "git abc list"
"git abc list" is using "git log" internally and printing in a format
similar to --pretty="%h %s", however sometimes it is useful to change
the option format slightly.  There is even already one example in
"git abc" itself in the form of "git pick-list".

Add options --oneline, --format and --pretty to "git list" and
when they are used invoke git log again to print each commit in the
requested format.  While at it add --reverse, which avoids using tac for
"git pick-list" but is also useful in general, e.g. with the default
--candidates list.