abc_flag: Don't warn when overwriting a flag with an identical flag
We added an extra warning when adding abc_flag_str, but didn't think
about the fact that we'd start warning on overwriting a flag with an
identical flag (which happens frequently when running reject on a list
of commits that may have already been rejected). Fix it by not updating
identical flags at all.

Fixes: c6dcd7f3 ("Add xcandidate flag and abc_flag_str")